Limiting the Number of Child Rays

+++ w1-3/src/color.rs
@@ -1,6 +1,6 @@
 use crate::vec3::Color;
 
-pub fn write_color(pixel_color: &Color, samples_per_pixel: i64) {
+pub fn write_color(pixel_color: &Color, samples_per_pixel: i32) {
   let scale = 1.0 / (samples_per_pixel as f64);
 
   let r = pixel_color.x * scale;
@@ -9,8 +9,8 @@ pub fn write_color(pixel_color: &Color, samples_per_pixel: i64) {
 
   println!(
     "{} {} {}",
-    (256.0 * r.clamp(0.0, 0.999)) as i64,
-    (256.0 * g.clamp(0.0, 0.999)) as i64,
-    (256.0 * b.clamp(0.0, 0.999)) as i64
+    (256.0 * r.clamp(0.0, 0.999)) as i32,
+    (256.0 * g.clamp(0.0, 0.999)) as i32,
+    (256.0 * b.clamp(0.0, 0.999)) as i32
   );
 }
